restructure: tier B — consolidate composed pipeline files
