+++ jetro-core/src/pipeline.rs
@@ -20,11 +20,7 @@ mod capability;
 mod collector;
 mod columnar;
 mod common;
-mod composed_barrier;
-mod composed_exec;
-mod composed_segment;
-mod composed_sink;
-mod composed_stage;
+mod composed;
 mod exec;
 mod indexed_exec;
 mod kernels;
