+++ jetro-core/src/pipeline/exec.rs
@@ -9,7 +9,7 @@ use crate::{
 };
 
 use super::columnar;
-use super::composed_exec;
+use super::composed;
 use super::indexed_exec;
 use super::materialized_exec;
 use super::{PhysicalExecPath, Pipeline, PipelineData};
@@ -42,7 +42,7 @@ impl Pipeline {
                 self.run_columnar_or_below(root, base_env, cache)
             }
             PhysicalExecPath::Columnar => self.run_columnar_or_below(root, base_env, cache),
-            PhysicalExecPath::Composed => composed_exec::run(self, root, base_env)
+            PhysicalExecPath::Composed => composed::run(self, root, base_env)
                 .unwrap_or_else(|| materialized_exec::run(self, root, base_env)),
             PhysicalExecPath::Legacy => materialized_exec::run(self, root, base_env),
         }
@@ -64,7 +64,7 @@ impl Pipeline {
                 return out;
             }
         }
-        composed_exec::run(self, root, base_env)
+        composed::run(self, root, base_env)
             .unwrap_or_else(|| materialized_exec::run(self, root, base_env))
     }
 }
